refactor: refactor result-category-detail-content folder structure
