+++ app/ui/result/result-category/result-category-detail-content/result-category-detail-content.stories.tsx
