+++ app/ui/result/result-category/result-category-detail-content/result-category-detail-content.tsx
@@ -2,8 +2,8 @@
 import { cn } from '@/app/utils/shadcn/utils';
 
 import { useState } from 'react';
-import { ResultCategoryDetailLectureToggle } from './result-category-detail-lecture-toggle';
-import ResultCagegoryDetailLecture from './result-cagegory-detail-lecture';
+import { ResultCategoryDetailLectureToggle } from '../result-category-detail-lecture-toggle';
+import ResultCagegoryDetailLecture from '../result-cagegory-detail-lecture';
 import { ResultCategoryDetailInfo } from '@/app/business/result/result.query';
 
 interface ResultCategoryDetailContentProps {
